+++ cfgov/v1/models/events.py
@@ -1,8 +1,8 @@
 import datetime
 
 from django.db import models
-from django.utils.functional import cached_property
-from django.template.response import TemplateResponse
+from django.db.models import Q
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 
 from localflavor.us.models import USStateField
 
@@ -11,6 +11,7 @@ from wagtail.wagtailadmin.edit_handlers import FieldPanel, MultiFieldPanel, Fiel
 from wagtail.wagtailimages.edit_handlers import ImageChooserPanel
 from wagtail.wagtaildocs.edit_handlers import DocumentChooserPanel
 from wagtail.wagtailcore import blocks
+from wagtail.wagtailcore.models import PAGE_TEMPLATE_VAR
 
 from .base import CFGOVPage
 
@@ -129,4 +130,59 @@ class EventPage(CFGOVPage):
 
 
 class EventLandingPage(CFGOVPage):
-    subpage_types = ['EventPage']
+    subpage_types = ['EventPage', 'EventArchivePage']
+
+    def get_context(self, request, *args, **kwargs):
+        return {
+            'events': EventPage.objects.live().filter(start_dt__gt=datetime.datetime.now()),
+            PAGE_TEMPLATE_VAR: self,
+            'self': self,
+            'request': request,
+        }
+
+# Importing here solves circular importing for now
+from v1.forms import EventsFilterForm
+
+
+class EventArchivePage(CFGOVPage):
+    body = models.TextField(blank=True)
+
+    subpage_types = []
+
+    content_panels = CFGOVPage.content_panels + [
+        FieldPanel('body', classname="full"),
+    ]
+
+    def get_context(self, request, *args, **kwargs):
+        form = EventsFilterForm(request.GET)
+        if form.is_valid():
+            # Generates a query by iterating over the zipped collection of
+            # tuples. The ordering of query strings is dependent on the
+            # ordering of the form fields in the definition of EventsFilterForm.
+            final_query = Q()
+            for field in zip(['start_dt__gte', 'end_dt__lte', 'tags__name__in'],
+                             form.fields):
+                if form.cleaned_data.get(field[1]):
+                    final_query &= \
+                        Q((field[0], form.cleaned_data.get(field[1])))
+
+        # List of paginated events to pass to the template
+        events_list = EventPage.objects.live().filter(final_query)
+        paginator = Paginator(events_list, 10)
+        page = request.GET.get('page')
+
+        try:
+            events = paginator.page(page)
+        except PageNotAnInteger:
+            events = paginator.page(1)
+        except EmptyPage:
+            events = paginator.page(paginator.num_pages)
+
+        return {
+            'form': form,
+            'filters': [field for field in form.fields],
+            'events': events,
+            PAGE_TEMPLATE_VAR: self,
+            'self': self,
+            'request': request,
+        }